Make State generic: remove unneeded typedb_service module
