+++ server/service/mod.rs
@@ -13,7 +13,6 @@ pub(crate) mod grpc;
 pub mod http;
 pub(crate) mod state;
 mod transaction_service;
-mod typedb_service;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialOrd, PartialEq, Eq, Hash)]
 #[serde(rename_all = "camelCase")]
